requirements.txt
fastapi==0.115.5
FastAPI_SQLAlchemy==0.2.1
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
gunicorn==21.2.0  # Add gunicorn to requirements
uvicorn==0.22.0  # Add uvicorn to your requirements
